refactor: change app structur
