+++ src/album/album.interface.ts
