+++ src/app.module.ts
@@ -1,10 +1,11 @@
 import { Module } from '@nestjs/common';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
-import { UserModule } from './modules/user.module';
+import { UserModule } from './users/user.module';
+import { TracksModule } from './tracks/tracks.module';
 
 @Module({
-  imports: [UserModule],
+  imports: [UserModule, TracksModule],
   controllers: [AppController],
   providers: [AppService],
 })
