+++ src/artist/artist.interface.ts
