+++ src/favorites/favorites.interface.ts
