+++ src/tracks/tracks.module.ts
@@ -1,0 +1,8 @@
+import { Module } from '@nestjs/common';
+
+@Module({
+  imports: [],
+  controllers: [],
+  providers: [],
+})
+export class TracksModule {}
