+++ src/types/appRoutes.interface.ts
